# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Spring Full-stack Starter API
  version: 1.0.0
  description: Todo API와 애플리케이션 상태 확인 API의 계약입니다.
servers:
  - url: http://localhost:8080
    description: Local Spring Boot server
paths:
  /actuator/health:
    get:
      operationId: getHealth
      summary: 애플리케이션 상태 조회
      responses:
        "200":
          description: 애플리케이션이 정상적으로 동작 중입니다.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/todos:
    get:
      operationId: listTodos
      summary: Todo 목록 조회
      parameters:
        - name: page
          in: query
          description: 0부터 시작하는 페이지 번호
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: 페이지당 항목 수
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 생성일 역순의 Todo 페이지
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoPageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      operationId: createTodo
      summary: Todo 생성
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: 생성된 Todo
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
  /api/todos/{id}:
    get:
      operationId: getTodo
      summary: Todo 단건 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 요청한 Todo
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: local 프로필에서만 사용하는 개발용 인증
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: prod 프로필에서 OIDC 발급자가 서명한 액세스 토큰
  headers:
    RequestId:
      description: 요청과 서버 로그를 연결하는 추적 ID
      schema:
        type: string
  responses:
    ValidationError:
      description: 요청값 검증 실패
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetail"
    AuthenticationRequired:
      description: 인증 정보가 없거나 올바르지 않음
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: 요청한 리소스가 존재하지 않음
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          examples: [UP]
      additionalProperties: true
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
    TodoResponse:
      type: object
      required: [id, title, completed, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoPageResponse:
      type: object
      required: [items, page, size, totalElements, totalPages, first, last]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TodoResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    ProblemDetail:
      type: object
      required: [title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
      additionalProperties: true
    ValidationProblemDetail:
      allOf:
        - $ref: "#/components/schemas/ProblemDetail"
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string